rock 띄우고 마우스 피킹으로 바운딩 구체와 충돌 체크 성공 -> 충돌 처리해줘야할듯 조명 쉐이더를 불린-총 쉐이딩 조명으로 변경 specular효과가 약간 바뀜

+++ game/Shader.cpp
@@ -179,8 +179,7 @@ GLint Shader::findLocation(const string& name) const
 // -------------------------------------------------------------------------
 void Shader::setBool(const std::string& name, bool value) const
 {
-	
-	glUniform1i(findLocation(name), (int)value);
+	glUniform1i(findLocation(name), (value));
 }
 // ------------------------------------------------------------------------
 void Shader::setInt(const std::string& name, int value) const
